+++ yax/core/src/main/scala/doobie/hi/resultset.scala
@@ -15,6 +15,7 @@ import doobie.free.{ databasemetadata => DMD }
 
 import doobie.util.composite._
 import doobie.util.invariant._
+import doobie.util.process.repeatEvalChunks
 
 import java.net.URL
 import java.util.{ Date, Calendar }
